Put Action messages into the CAS.
Experience has shown that having access to a copy of the Action message
available makes debugging much easier, as not doing so risks the client
simply throwing it away. It also results in duplication, with the same
or similar Actions being sent repeatedly, wasting bandwidth.

Putting the Action messages into the CAS ensure that they will remain
accessible, and additionally allows servers to perform access control on
UpdateActionResult based on the contents of an Action (for instance, to
prevent an individual user from poisoning actions that might be executed
by a continuous integration system).