Document that parent directories of outputs are created implicitly.
When implementing Buildbarn's worker, I noticed that there exists the
assumption in build rules that directories under bazel-out/ are already
created prior to execution, even if they are not part of the input root.

As I suspect that getting rid of this assumption is both unrealistic and
undesirable, let's document this instead.

Mentioned in: #40