Document what the intended file permissions of the input root is.
At the build event at Bloomberg, several users of REv2 (Bazel,
Buildbarn, BuildGrid) discussed what file permissions of the input root
ought to be. Basically, there were two camps:

- Camp A: reduce the number of writable files and directories to the
  absolute minumum (Bazel and Buildbarn).

- Camp B: keep files read-only to allow caching based on hardlinks,
  while making all directories writable (BuildGrid). This is desired to
  support the execution of very large build actions that invoke
  traditional build systems that tend to clobber the input root all over
  the place.

It was eventually decided to support the latter. In the long run a
separate effort could be started to lock this down further, if it turns
out this causes significant performance overhead.

Until recently, Buildbarn also followed the second approach. A recent
change caused Buildbarn to only make directories with one or more
outputs to be writable and all of the parent directories. This change
will be reverted to match the semantics proposed by this commit.

Fixes: #43