support inlining of output files in the action result
Allow the client to request the inlining of action outputs,
stdout and stderr in the action result message. We only add
this capability to the GetActionResult call for now as its
typically on the latency sensitive fast path, while Execute
is not.

Also make the stdout and stderr inlining decisions client
controlled and recommend for the server to keep the serialized
size of the ActionResult message below 1 MiB [1].

See https://groups.google.com/d/msg/remote-execution-apis/bmUuI2wUcwA/p0yrYq7VBgAJ

[1] https://developers.google.com/protocol-buffers/docs/techniques#large-data