Field-based node properties (#128)
This reduces the size of directory trees compared to string-based node
properties by allowing node properties to be defined as regular protobuf
fields in the new message type `NodeProperties`.

The existing string-based node properties are also moved into the new
message type to avoid having to define the same two fields
(`repeated NodeProperty` and the new `NodeProperties`) in various places
in the proto (and derivative protos). This also allows client/server
code to handle all node properties with a single message object.