point out that servers may modify ActionResults, and mention one reason why
When serialized, default ActionResult proto3 messages contain no data.
This is unfortunate because there is no way to distinguish this valid data
from from file truncation errors on the server (which are perhaps the most
common form of file-based data corruption).

As a workaround, server implementations can modify incoming ActionResult
messages to ensure that at least one field has a non-default value, and then
consider any default/zero-length ActionResult messages retrieved from
storage as corrupt. Let's point this out in the documentation.

The simplest field that servers may consider modifying is
ActionResult.execution_metadata.worker, which (if empty) can safely be set
to the IP address or hostname of the creator/uploader without modifying the
semantic meaning of the ActionResult.