[ci] autogenerate go bindings on master
This is a github action which attempts to keep the generated *.pb.go
code in sync with the corresponding .proto files on master. This action
runs when master is updated- it takes a few minutes to run and will
push a new commit to master if there were any changes required.

I assume if there are more than one of these actions running at the same
time and both try to push, only the first will succeed. We should try to
avoid this situation by not landing changes until the previous action has
completed.