Document that Execute() is permitted to run actions more than once
As discussed during yesterday's Remote Execution API Working Group
meeting, there may be cases in which it's desirable to run actions more
than once. For example, a cluster could first try to run the action on a
small worker, falling back to a larger worker in case of failure.

Let's make it explicit that the Execute() call doesn't guarantee that an
action is only run once. This makes it clear that the Remote Execution
protocol isn't suitable for running tasks that aren't idempotent.