Add the Remote Output API
As being discussed in this thread, I am currently working on letting
Bazel store its output directory on a FUSE file system:

https://groups.google.com/g/remote-execution-apis/c/qOSWWwBLPzo/m/TbcNYzwMBwAJ

The plan is to let a separate daemon manage the FUSE daemon. It's not
going to be part of Bazel itself. This is why I've worked on a gRPC
protocol that Bazel can use to send instructions to the FUSE daemon.
This commit brings in a first version of this protocol.