Add support for interleaved stdout/stderr outputs
REv2 currently stores stdout/stderr as two separate objects. A downside
of this approach is that information gets lost on how log messages are
ordered between these two streams. Another issue is that no timestamp
information is stored. When tests fail due to timeouts, it may be useful
to figure out which parts of the test take long to run.

This proposal is an alternative to PR #173. The main difference is that
it defines a Protobuf message whose structure is identical to the
messages generated by Docker's JSON logging module.

https://docs.docker.com/config/containers/logging/json-file/