Add support for the Deflate compression algorithm
I don't think that it's a good idea to use Deflate in general. That
said, the reason I'm proposing that we add it is because it allows a
remote execution service to do light-weight "extraction" of ZIP files.
An existing ZIP file can be carved up into individual file payloads,
without decompressing/recompressing them.

Furthermore, adding this algorithm will allow Buildbarn to get rid of a
similar enum, which already includes Deflate.