Add support for inlined compressed data (larger refactoring)
By defining a Blob message that overlaps with Digest, we can repurpose
Digest fields to allow data inlining with optional compression in more
places.  The Digest message is then only used in cases where the
receiver wants a reference rather than data.

In this example I have changed most of the Digest fields to Blobs,
which may lead to over-inlining of data.  Each case needs to be
considered separately before this change is ready for consideration.

Refers to #201.