Relax input tree requirements
To reduce the workload of the remote execution clients, the requirement
of constructing Merkle trees is removed. Instead, the client can choose
any structure of it input file tree it would like. A client might choose
to replicate the internal dependency graph.

The down side is that different clients, even different versions of the
clients, might not get cache hit on effectively the same action. The use
case of sharing the cache for different clients is not a realistic use
case. Also, clients are most likely stable in their serialization
between versions.

Fixes #141.