Make it explicit that the timeout may not be in real time (#204)
Buildbarn's FUSE file system is capable of lazily loading the input root
of an action as it's being executed. Because this adds a
non-deterministic amount of noise to execution times, there is an option
for automatically compensating the execution timeout of time spent
reading data from the CAS. Let's extend the spec to explicitly allow this.

While there, add a new field to ExecutedActionMetadata that contains the
execution time of the action, using a method of timekeeping that is
consistent with the timeout.