remote_asset: link digest_function & checksum qual
Ensure that the digest function in requests is consistent with the
digest function used in `checksum.sri` qualifier, this way server
implementations do not have to support replicating blob from one digest
function to another.

Such functionality is not the intended use case of Remote Asset APIs
and, instead, could be achieved via a different client-side API call.