RequestMetadata: Make tool_details repeated
It is possible for more than one tool to be involved in the sending of a
gRPC request, for example when a client-side proxy is used it is
interesting to know the name and version of both the proxy and the
original client tool.

To facilitate this in RequestMetadata, this commit updates the
tool_details field to be `repeated`, allowing proxies and other related
tools to also be recorded in the metadata sent with requests.