RequestMetadata: Add additional_handlers field
It is possible for more than one tool to be involved in the sending of a
gRPC request, for example when a client-side proxy is used it is
interesting to know the name and version of both the proxy and the
original client tool.

To facilitate this in RequestMetadata, this commit adds a new repeated
`additional_handlers` field to the message which can contain details of
other tools involved in the handling of a request.