Add max_blob_size_bytes to CacheCapabilities
Some CAS implementations place an upper limit on the size of blobs that
they will accept in uploads. There is currently no standard way for clients
to query cache servers for this limit, and rejected large uploads can be
slow.

So let's add this to the CacheCapabilities API. For backwards compatibility,
a value of 0 means "no specified limit" (ie not "no uploads allowed").