asset API: loosen restrictions on qualifiers
The spec previously stated that servers must reject Fetch requests if
there were unsupported qualifiers specified. This was too strict- consider
for example a scenario where the client specifies both SHA256 and SHA512
qualifiers, and the server only supports one of these. Supporting only one
of these should be sufficient, and the server should be free to accept
such a request.