CacheCapabilities.max_batch_total_size_bytes refers to the size of the serialized protobuf messages, not the blobs
I believe that this value was intended to help clients and servers
avoid exceeding gRPC message length limits, which default to 4Mb in
most configurations.

When dealing with uncompressed blobs, the total size of the blobs is
often approximately equal to the size of the serialized protobuf message.
However this approximation is much less reliable when compressed blobs
are involved in response messages (where the client doesn't know ahead
of time how large the compressed blobs are).

We should therefore describe this limit as referring to the size of
the serialized requests/responses, and not of the approximation using
the sizes of the blobs (either uncompressed or compressed).

Resolves #344.